Moved all PaginationComponents into the new Pagination folder
